requirements.txt
streamlit>=1.31.0
pydantic>=2.6.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=4.0.0
